+++ R/availability.R
@@ -21,10 +21,10 @@
 
 #' Check availability of columns
 #'
-#' Easy check for availability of columns in a data set. This makes it easy to get an idea of which antimicrobial combination can be used for calculation with e.g. \code{\link{portion_R}}.
+#' Easy check for availability of columns in a data set. This makes it easy to get an idea of which antimicrobial combination can be used for calculation with e.g. \code{\link{resistance}}.
 #' @param tbl a \code{data.frame} or \code{list}
 #' @param width number of characters to present the visual availability, defaults to filling the width of the console
-#' @details The function returns a \code{data.frame} with columns \code{"resistant"} and \code{"visual_resistance"}. The values in that columns are calculated with \code{\link{portion_R}}.
+#' @details The function returns a \code{data.frame} with columns \code{"resistant"} and \code{"visual_resistance"}. The values in that columns are calculated with \code{\link{resistance}}.
 #' @return \code{data.frame} with column names of \code{tbl} as row names
 #' @inheritSection AMR Read more on our website!
 #' @importFrom cleaner percentage
@@ -48,7 +48,7 @@ availability <- function(tbl, width = NULL) {
     1 - base::sum(base::is.na(x)) / base::length(x) 
   })
   n <- base::sapply(tbl, function(x) base::length(x[!base::is.na(x)]))
-  R <- base::sapply(tbl, function(x) base::ifelse(is.rsi(x), portion_R(x, minimum = 0), NA))
+  R <- base::sapply(tbl, function(x) base::ifelse(is.rsi(x), resistance(x, minimum = 0), NA))
   R_print <- character(length(R))
   R_print[!is.na(R)] <- percentage(R[!is.na(R)])
   R_print[is.na(R)] <- ""