+++ R/count.R
@@ -23,20 +23,19 @@
 #'
 #' @description These functions can be used to count resistant/susceptible microbial isolates. All functions support quasiquotation with pipes, can be used in \code{dplyr}s \code{\link[dplyr]{summarise}} and support grouped variables, see \emph{Examples}.
 #'
-#' \code{count_R} and \code{count_IR} can be used to count resistant isolates, \code{count_S} and \code{count_SI} can be used to count susceptible isolates.\cr
+#' \code{count_resistant} should be used to count resistant isolates, \code{count_susceptible} should be used to count susceptible isolates.\cr
 #' @param ... one or more vectors (or columns) with antibiotic interpretations. They will be transformed internally with \code{\link{as.rsi}} if needed.
-#' @inheritParams portion
+#' @inheritParams proportion
 #' @inheritSection as.rsi Interpretation of S, I and R
-#' @details These functions are meant to count isolates. Use the \code{\link{portion}_*} functions to calculate microbial resistance.
+#' @details These functions are meant to count isolates. Use the \code{\link{resistance}}/\code{\link{susceptibility}} functions to calculate microbial resistance/susceptibility.
+#' 
+#' The function \code{count_resistant} is equal to the function \code{count_R}. The function \code{count_susceptible} is equal to the function \code{count_SI}.
 #'
-#' The function \code{n_rsi} is an alias of \code{count_all}. They can be used to count all available isolates, i.e. where all input antibiotics have an available result (S, I or R). Their use is equal to \code{\link{n_distinct}}. Their function is equal to \code{count_S(...) + count_IR(...)}.
+#' The function \code{n_rsi} is an alias of \code{count_all}. They can be used to count all available isolates, i.e. where all input antibiotics have an available result (S, I or R). Their use is equal to \code{\link{n_distinct}}. Their function is equal to \code{count_susceptible(...) + count_resistant(...)}.
 #'
-#' The function \code{count_df} takes any variable from \code{data} that has an \code{"rsi"} class (created with \code{\link{as.rsi}}) and counts the amounts of S, I and R. The resulting \emph{tidy data} (see Source) \code{data.frame} will have three rows (S/I/R) and a column for each variable with class \code{"rsi"}.
-#'
-#' The function \code{rsi_df} works exactly like \code{count_df}, but adds the percentage of S, I and R.
-#' @inheritSection portion Combination therapy
-#' @source Wickham H. \strong{Tidy Data.} The Journal of Statistical Software, vol. 59, 2014. \url{http://vita.had.co.nz/papers/tidy-data.html}
-#' @seealso \code{\link{portion}_*} to calculate microbial resistance and susceptibility.
+#' The function \code{count_df} takes any variable from \code{data} that has an \code{"rsi"} class (created with \code{\link{as.rsi}}) and counts the number of S's, I's and R's. The function \code{rsi_df} works exactly like \code{count_df}, but adds the percentage of S, I and R.
+#' @inheritSection proportion Combination therapy
+#' @seealso \code{\link{proportion}_*} to calculate microbial resistance and susceptibility.
 #' @return Integer
 #' @rdname count
 #' @name count
@@ -45,24 +44,28 @@
 #' @examples
 #' # example_isolates is a data set available in the AMR package.
 #' ?example_isolates
+#' 
+#' count_resistant(example_isolates$AMX)   # counts "R"
+#' count_susceptible(example_isolates$AMX) # counts "S" and "I"
+#' count_all(example_isolates$AMX)         # counts "S", "I" and "R"
 #'
-#' # Count resistant isolates
-#' count_R(example_isolates$AMX)
-#' count_IR(example_isolates$AMX)
-#'
-#' # Or susceptible isolates
+#' # be more specific
 #' count_S(example_isolates$AMX)
 #' count_SI(example_isolates$AMX)
+#' count_I(example_isolates$AMX)
+#' count_IR(example_isolates$AMX)
+#' count_R(example_isolates$AMX)
 #'
 #' # Count all available isolates
 #' count_all(example_isolates$AMX)
 #' n_rsi(example_isolates$AMX)
 #'
-#' # Since n_rsi counts available isolates, you can
-#' # calculate back to count e.g. non-susceptible isolates.
-#' # This results in the same:
-#' count_SI(example_isolates$AMX)
-#' portion_SI(example_isolates$AMX) * n_rsi(example_isolates$AMX)
+#' # n_rsi() is an alias of count_all().
+#' # Since it counts all available isolates, you can
+#' # calculate back to count e.g. susceptible isolates.
+#' # These results are the same:
+#' count_susceptible(example_isolates$AMX)
+#' susceptibility(example_isolates$AMX) * n_rsi(example_isolates$AMX)
 #'
 #' library(dplyr)
 #' example_isolates %>%
@@ -76,19 +79,17 @@
 #'
 #' # Count co-resistance between amoxicillin/clav acid and gentamicin,
 #' # so we can see that combination therapy does a lot more than mono therapy.
-#' # Please mind that `portion_SI` calculates percentages right away instead.
-#' count_SI(example_isolates$AMC)  # 1433
-#' count_all(example_isolates$AMC) # 1879
+#' # Please mind that `susceptibility()` calculates percentages right away instead.
+#' example_isolates %>% count_susceptible(AMC) # 1433
+#' example_isolates %>% count_all(AMC)         # 1879
 #'
-#' count_SI(example_isolates$GEN)  # 1399
-#' count_all(example_isolates$GEN) # 1855
+#' example_isolates %>% count_susceptible(GEN) # 1399
+#' example_isolates %>% count_all(GEN)         # 1855
 #'
-#' with(example_isolates,
-#'      count_SI(AMC, GEN))        # 1764
-#' with(example_isolates,
-#'      n_rsi(AMC, GEN))           # 1936
-#'
-#' # Get portions S/I/R immediately of all rsi columns
+#' example_isolates %>% count_susceptible(AMC, GEN) # 1764
+#' example_isolates %>% count_all(AMC, GEN)         # 1936
+
+#' # Get number of S+I vs. R immediately of selected columns
 #' example_isolates %>%
 #'   select(AMX, CIP) %>%
 #'   count_df(translate = FALSE)
@@ -99,6 +100,24 @@
 #'   group_by(hospital_id) %>%
 #'   count_df(translate = FALSE)
 #'
+count_resistant <- function(..., only_all_tested = FALSE) {
+  rsi_calc(...,
+           ab_result = "R",
+           only_all_tested = only_all_tested,
+           only_count = TRUE)
+}
+
+#' @rdname count
+#' @export
+count_susceptible <- function(..., only_all_tested = FALSE) {
+  rsi_calc(...,
+           ab_result = c("S", "I"),
+           only_all_tested = only_all_tested,
+           only_count = TRUE)
+}
+
+#' @rdname count
+#' @export
 count_R <- function(..., only_all_tested = FALSE) {
   rsi_calc(...,
            ab_result = "R",
@@ -109,6 +128,7 @@ count_R <- function(..., only_all_tested = FALSE) {
 #' @rdname count
 #' @export
 count_IR <- function(..., only_all_tested = FALSE) {
+  warning("Using 'count_IR' is discouraged; use 'count_resistant()' instead to not consider \"I\" being resistant.", call. = FALSE)
   rsi_calc(...,
            ab_result = c("I", "R"),
            only_all_tested = only_all_tested,
@@ -136,6 +156,7 @@ count_SI <- function(..., only_all_tested = FALSE) {
 #' @rdname count
 #' @export
 count_S <- function(..., only_all_tested = FALSE) {
+  warning("Using 'count_S' is discouraged; use 'count_susceptible()' instead to also consider \"I\" being susceptible.", call. = FALSE)
   rsi_calc(...,
            ab_result = "S",
            only_all_tested = only_all_tested,