+++ R/rsi_calc.R
@@ -22,7 +22,7 @@
 #' @importFrom rlang enquos as_label
 dots2vars <- function(...) {
   # this function is to give more informative output about 
-  # variable names in count_* and portion_* functions
+  # variable names in count_* and proportion_* functions
   paste(
     unlist(
       lapply(enquos(...),
@@ -46,9 +46,9 @@ rsi_calc <- function(...,
                      as_percent = FALSE,
                      only_all_tested = FALSE,
                      only_count = FALSE) {
-
+  
   data_vars <- dots2vars(...)
-
+  
   if (!is.numeric(minimum)) {
     stop("`minimum` must be numeric", call. = FALSE)
   }
@@ -58,47 +58,47 @@ rsi_calc <- function(...,
   if (!is.logical(only_all_tested)) {
     stop("`only_all_tested` must be logical", call. = FALSE)
   }
-
+  
   dots_df <- ...elt(1) # it needs this evaluation
   dots <- base::eval(base::substitute(base::alist(...)))
   if ("also_single_tested" %in% names(dots)) {
-    stop("`also_single_tested` was replaced by `only_all_tested`. Please read Details in the help page (`?portion`) as this may have a considerable impact on your analysis.", call. = FALSE)
+    stop("`also_single_tested` was replaced by `only_all_tested`. Please read Details in the help page (`?proportion`) as this may have a considerable impact on your analysis.", call. = FALSE)
   }
   ndots <- length(dots)
-
- if ("data.frame" %in% class(dots_df)) {
-   # data.frame passed with other columns, like: example_isolates %>% portion_S(amcl, gent)
-   dots <- as.character(dots)
-   dots <- dots[dots != "."]
+  
+  if ("data.frame" %in% class(dots_df)) {
+    # data.frame passed with other columns, like: example_isolates %>% proportion_S(amcl, gent)
+    dots <- as.character(dots)
+    dots <- dots[dots != "."]
     if (length(dots) == 0 | all(dots == "df")) {
-      # for complete data.frames, like example_isolates %>% select(amcl, gent) %>% portion_S()
+      # for complete data.frames, like example_isolates %>% select(amcl, gent) %>% proportion_S()
       # and the old rsi function, that has "df" as name of the first parameter
       x <- dots_df
     } else {
       x <- dots_df[, dots]
     }
   } else if (ndots == 1) {
-    # only 1 variable passed (can also be data.frame), like: portion_S(example_isolates$amcl) and example_isolates$amcl %>% portion_S()
+    # only 1 variable passed (can also be data.frame), like: proportion_S(example_isolates$amcl) and example_isolates$amcl %>% proportion_S()
     x <- dots_df
   } else {
-    # multiple variables passed without pipe, like: portion_S(example_isolates$amcl, example_isolates$gent)
+    # multiple variables passed without pipe, like: proportion_S(example_isolates$amcl, example_isolates$gent)
     x <- NULL
     try(x <- as.data.frame(dots), silent = TRUE)
     if (is.null(x)) {
-      # support for: with(example_isolates, portion_S(amcl, gent))
+      # support for: with(example_isolates, proportion_S(amcl, gent))
       x <- as.data.frame(rlang::list2(...))
     }
   }
-
+  
   if (is.null(x)) {
     warning("argument is NULL (check if columns exist): returning NA", call. = FALSE)
     return(NA)
   }
-
+  
   print_warning <- FALSE
-
+  
   ab_result <- as.rsi(ab_result)
-
+  
   if (is.data.frame(x)) {
     rsi_integrity_check <- character(0)
     for (i in seq_len(ncol(x))) {
@@ -113,7 +113,7 @@ rsi_calc <- function(...,
       # this will give a warning for invalid results, of all input columns (so only 1 warning)
       rsi_integrity_check <- as.rsi(rsi_integrity_check)
     }
-
+    
     if (only_all_tested == TRUE) {
       # THE NUMBER OF ISOLATES WHERE *ALL* ABx ARE S/I/R
       x <- apply(X = x %>% mutate_all(as.integer),
@@ -140,16 +140,16 @@ rsi_calc <- function(...,
     numerator <- sum(x %in% ab_result, na.rm = TRUE)
     denominator <- sum(x %in% levels(ab_result), na.rm = TRUE)
   }
-
+  
   if (print_warning == TRUE) {
     warning("Increase speed by transforming to class `rsi` on beforehand: df %>% mutate_if(is.rsi.eligible, as.rsi)",
             call. = FALSE)
   }
-
+  
   if (only_count == TRUE) {
     return(numerator)
   }
-
+  
   if (denominator < minimum) {
     if (data_vars != "") {
       data_vars <- paste(" for", data_vars)
@@ -159,7 +159,7 @@ rsi_calc <- function(...,
   } else {
     fraction <- numerator / denominator
   }
-
+  
   if (as_percent == TRUE) {
     percentage(fraction, digits = 1)
   } else {
@@ -169,7 +169,7 @@ rsi_calc <- function(...,
 
 #' @importFrom dplyr %>% summarise_if mutate select everything bind_rows
 #' @importFrom tidyr gather
-rsi_calc_df <- function(type, # "portion" or "count"
+rsi_calc_df <- function(type, # "proportion" or "count"
                         data,
                         translate_ab = "name",
                         language = get_locale(),
@@ -178,79 +178,81 @@ rsi_calc_df <- function(type, # "portion" or "count"
                         combine_SI = TRUE,
                         combine_IR = FALSE,
                         combine_SI_missing = FALSE) {
-
+  
   if (!"data.frame" %in% class(data)) {
     stop(paste0("`", type, "_df` must be called on a data.frame"), call. = FALSE)
   }
-
+  
   if (isTRUE(combine_IR) & isTRUE(combine_SI_missing)) {
     combine_SI <- FALSE
   }
   if (isTRUE(combine_SI) & isTRUE(combine_IR)) {
     stop("either `combine_SI` or `combine_IR` can be TRUE, not both", call. = FALSE)
   }
-
+  
   if (!any(sapply(data, is.rsi), na.rm = TRUE)) {
     stop("No columns with class 'rsi' found. See ?as.rsi.", call. = FALSE)
   }
-
+  
   if (as.character(translate_ab) %in% c("TRUE", "official")) {
     translate_ab <- "name"
   }
-
+  
   get_summaryfunction <- function(int, type) {
-    # look for portion_S, count_S, etc:
+    # look for proportion_S, count_S, etc:
     int_fn <- get(paste0(type, "_", int), envir = asNamespace("AMR"))
-
-    if (type == "portion") {
-      summ <- summarise_if(.tbl = data,
-                           .predicate = is.rsi,
-                           .funs = int_fn,
-                           minimum = minimum,
-                           as_percent = as_percent)
-    } else if (type == "count") {
-      summ <- summarise_if(.tbl = data,
-                           .predicate = is.rsi,
-                           .funs = int_fn)
-    }
+    
+    suppressWarnings(
+      if (type == "proportion") {
+        summ <- summarise_if(.tbl = data,
+                             .predicate = is.rsi,
+                             .funs = int_fn,
+                             minimum = minimum,
+                             as_percent = as_percent)
+      } else if (type == "count") {
+        summ <- summarise_if(.tbl = data,
+                             .predicate = is.rsi,
+                             .funs = int_fn)
+      }
+    )
     summ %>%
       mutate(interpretation = int) %>%
       select(interpretation, everything())
   }
-
+  
   resS <- get_summaryfunction("S", type)
   resI <- get_summaryfunction("I", type)
   resR <- get_summaryfunction("R", type)
   resSI <- get_summaryfunction("SI", type)
   resIR <- get_summaryfunction("IR", type)
   data.groups <- group_vars(data)
-
+  
   if (isFALSE(combine_SI) & isFALSE(combine_IR)) {
     res <- bind_rows(resS, resI, resR) %>%
       mutate(interpretation = factor(interpretation,
                                      levels = c("S", "I", "R"),
                                      ordered = TRUE))
-
+    
   } else if (isTRUE(combine_IR)) {
     res <- bind_rows(resS, resIR) %>%
       mutate(interpretation = factor(interpretation,
                                      levels = c("S", "IR"),
                                      ordered = TRUE))
-
+    
   } else if (isTRUE(combine_SI)) {
     res <- bind_rows(resSI, resR) %>%
       mutate(interpretation = factor(interpretation,
                                      levels = c("SI", "R"),
                                      ordered = TRUE))
   }
-
+  
   res <- res %>%
     gather(antibiotic, value, -interpretation, -data.groups) %>%
     select(antibiotic, everything())
-
+  
   if (!translate_ab == FALSE) {
     res <- res %>% mutate(antibiotic = AMR::ab_property(antibiotic, property = translate_ab, language = language))
   }
-
+  
   res
 }