+++ NEWS.md
@@ -1,17 +1,20 @@
-# AMR 0.8.0.9026
+# AMR 0.8.0.9027
 <small>Last updated: 10-Nov-2019</small>
 
 ### New
+* Functions `susceptibility()` and `resistance()` as aliases of `proportion_SI()` and `proportion_R()`, respectively. These functions were added to make it more clear that I should be considered susceptible and not resistant.
 * Support for a new MDRO guideline: Magiorakos AP, Srinivasan A *et al.* "Multidrug-resistant, extensively drug-resistant and pandrug-resistant bacteria: an international expert proposal for interim standard definitions for acquired resistance." Clinical Microbiology and Infection (2012).
   * This is now the new default guideline for the `mdro()` function
   * The new Verbose mode (`mdro(...., verbose = TRUE)`) returns an informative data set where the reason for MDRO determination is given for every isolate, and an list of the resistant antimicrobial agents
 
 ### Changes
+* Removed previously deprecated function `as.rsi()` - this function was replaced by `ab_atc()`
+* Renamed all `portion_*` functions to `proportion_*`. All `portion_*` functions are still available as deprecated functions, and will return a warning when used.
 * When running `as.rsi()` over a data set, it will now print the guideline that will be used if it is not specified by the user
 * Fix for `eucast_rules()`: *Stenotrophomonas maltophilia* not interpreted "R" to ceftazidime anymore (following EUCAST v3.1)
 * Adopted Adeolu *et al.* (2016), [PMID 27620848](https://www.ncbi.nlm.nih.gov/pubmed/27620848) for the `microorganisms` data set, which means that the new order Enterobacterales now consists of a part of the existing family *Enterobacteriaceae*, but that this family has been split into other families as well (like *Morganellaceae* and *Yersiniaceae*). Although published in 2016, this information is not yet in the Catalogue of Life version of 2019. All MDRO determinations with `mdro()` will now use the Enterobacterales order for all guidelines before 2016.
 * Fix for interpreting MIC values with `as.rsi()` where the input is `NA`
-* Added "imi" as allowed abbreviation for Imipenem
+* Added "imi" and "imp" as allowed abbreviation for Imipenem (IPM)
 * Fix for automatically determining columns with antibiotic results in `mdro()` and `eucast_rules()`
 * Added ATC codes for ceftaroline, ceftobiprole and faropenem and fixed two typos in the `antibiotics` data set
 * More robust way of determining valid MIC values